levers: split the tests out to stay under the file-size cap
The five SSM fields and their polarity assertions pushed `model_levers.rs` to
523 lines, over the repository's 500-LoC cap and not allow-listed — I pushed
before checking, so this is the follow-up. Tests move to
`model_levers_tests.rs` behind `#[path]`, the pattern `mtp_carry_tests.rs`
already uses: 283 and 250 lines.

★ The `from_env` guard then failed on its own new home. It greps `crates/` for
callers of `ModelLevers::from_env()`, and a test that forbids a string
necessarily contains it — so moving the tests into their own file made that
file an offender. Allow-listed with the reason. Worth keeping as a note: the
guard noticed a file move it was never told about, which is what a source-level
check is for.
